Moved connection monitor to dedicated file.
